+++ src/main/scala/ckite/rpc/RequestVote.scala
@@ -1,5 +1,5 @@
 package ckite.rpc
 
-case class RequestVote(memberId: String, term: Int, lastLogIndex: Int = -1, lastLogTerm: Int = -1) {
+case class RequestVote(memberId: String, term: Int, lastLogIndex: Long = -1, lastLogTerm: Int = -1) {
   override def toString():String = s"RequestVote($memberId,term=$term,lastLogIndex=$lastLogIndex,lastLogTerm=$lastLogTerm)"
 }
